refactor load liquid command getter
