+++ protocol-designer/src/load-file/migration/6_0_0.ts
@@ -12,9 +12,7 @@ import { uuid } from '../../utils'
 //  and labware access parameters, renames AirGap to aspirate, and removes all temporal properties from labware, pipettes,
 //  and module keys such as slot, mount
 //  and renames well to wellName
-import {
-  LoadLiquidCreateCommand
-} from '../../../../shared-data/protocol/types/schemaV6/command/setup';
+import { getLoadLiquidCommands } from './utils/getLoadLiquidCommands';
 import type {
   LoadPipetteCreateCommand,
   LoadModuleCreateCommand,
@@ -24,19 +22,7 @@ import type {
   CreateCommand,
   ProtocolFile,
 } from '@opentrons/shared-data/protocol/types/schemaV6'
-
-interface DesignerApplicationData {
-  ingredients: Record<
-    string,
-    {
-      name: string
-      description: string | null
-      serialize: boolean
-      liquidGroupId: string
-    }
-  >
-  ingredLocations: { [labwareId: string]: { [wellName: string]: { [liquidId: string]: { volume: number } } } }
-}
+import type { DesignerApplicationData } from './utils/getLoadLiquidCommands';
 
 const PD_VERSION = '6.0.0'
 const SCHEMA_VERSION = 6
@@ -135,63 +121,7 @@ export const migrateFile = (
     }
   )
 
-  let loadLiquidCommands: LoadLiquidCreateCommand[] = []
-
-  let labwareIdsByLiquidId: { [liquidId: string]: string[] } = {}
-
-  if (designerApplication?.data?.ingredLocations != null && designerApplication?.data?.ingredients != null) {
-    Object.keys(designerApplication?.data?.ingredients).forEach(liquidId => {
-      if (designerApplication?.data?.ingredLocations != null) {
-        for (const [labwareId, liquidsByWellName] of Object.entries(designerApplication?.data?.ingredLocations)) {
-          // eslint-disable-next-line @typescript-eslint/no-unused-vars
-          for (const [_wellName, volumeByLiquidId] of Object.entries(liquidsByWellName)) {
-            if (liquidId in volumeByLiquidId) {
-              if (labwareIdsByLiquidId[liquidId] == null) {
-                labwareIdsByLiquidId = {
-                  ...labwareIdsByLiquidId,
-                  [liquidId]: [labwareId]
-                }
-              } else if (!labwareIdsByLiquidId[liquidId].includes(labwareId)) {
-                labwareIdsByLiquidId = {
-                  ...labwareIdsByLiquidId,
-                  [liquidId]: [...labwareIdsByLiquidId[liquidId], labwareId]
-                }
-              }
-            }
-          }
-        }
-      }
-    })
-
-    loadLiquidCommands = reduce<{ [liquidId: string]: string[] }, LoadLiquidCreateCommand[]>(labwareIdsByLiquidId, (acc, labwareIds, liquidId) => {
-      const commands: LoadLiquidCreateCommand[] = labwareIds.map(labwareId => {
-        const volumeByWell = reduce(designerApplication.data?.ingredLocations[labwareId], (acc, volumesByLiquidId, wellName) => {
-          if (liquidId in volumesByLiquidId) {
-            return {
-              ...acc,
-              [wellName]: volumesByLiquidId[liquidId].volume
-            }
-          } else {
-            return { ...acc }
-          }
-        }, {})
-
-        const loadLiquidCommand: LoadLiquidCreateCommand = {
-          commandType: "loadLiquid",
-          key: uuid(),
-          params: {
-            liquidId,
-            labwareId,
-            volumeByWell
-          }
-        }
-        return loadLiquidCommand
-      })
-      return [...commands, ...acc]
-
-    }, [])
-  }
-
+  const loadLiquidCommands = getLoadLiquidCommands(designerApplication)
   const migratedV5Commands = migrateCommands(commands)
 
   const liquids: ProtocolFile['liquids'] =
